requirements.txt
#asgiref==3.8.1
#Django==5.1.1
#django-cors-headers==4.4.0
#djangorestframework==3.15.2
#djangorestframework-simplejwt==5.3.1
#psycopg2-binary==2.9.9
#PyJWT==2.9.0
#pytz==2024.2
#sqlparse==0.5.1
#tzdata==2024.1


Django==4.0.3
djangorestframework
django-cors-headers
djangorestframework-simplejwt
psycopg2-binary
pandas
openpyxl
xlsxwriter
requests
Jinja2
xhtml2pdf
Office365-REST-Python-Client
reportlab
weasyprint
num2words
mysqlclient

